refactor: 패키지 생성 및 이동 #8
기존 dto 패키지 삭제,
entity 하위에 response, request 패키지 생성 및 관련 클래스 이동
