+++ be/src/main/java/buddyguard/mybuddyguard/weight/contoller/response/WeightResponse.java
@@ -1,4 +1,4 @@
-package buddyguard.mybuddyguard.weight.dto;
+package buddyguard.mybuddyguard.weight.contoller.response;
 
 import java.time.LocalDateTime;
